+++ packages/plugin-packages/mashroom-portal-demo-vue-app/src/js/index.js
@@ -1,30 +1,24 @@
 
-import Vue from 'vue';
+import { createApp } from 'vue';
 import App from './App.vue';
 
 const bootstrap = (portalAppHostElement, portalAppSetup, clientServices) => {
     const { resourcesBasePath, appConfig: { message, pingButtonLabel }} = portalAppSetup;
     const { messageBus } = clientServices;
 
-    // Vue seems to replace the host element, so we create a sub node
-    portalAppHostElement.innerHTML = '<div />';
-
-    const vm = new Vue({
-        el: portalAppHostElement.firstElementChild,
-        render: h => h(App, {
-            props: {
-                resourcesBasePath,
-                message,
-                pingButtonLabel,
-                messageBus
-            }
-        })
+    const app = createApp(App, {
+        resourcesBasePath,
+        message,
+        pingButtonLabel,
+        messageBus
     });
+
+    app.mount(portalAppHostElement);
 
     return {
         willBeRemoved: () => {
             console.info('Destroying Vue app');
-            vm.$destroy();
+            app.unmount();
         }
     };
 };