+++ packages/plugin-packages/mashroom-portal-demo-vue-app/src/js/indexSSR.js
@@ -1,0 +1,20 @@
+
+import { createSSRApp } from 'vue';
+import { renderToString } from 'vue/server-renderer'
+import App from './App';
+
+const bootstrap = async (portalAppSetup) => {
+    const { resourcesBasePath, appConfig: { message, pingButtonLabel }} = portalAppSetup;
+    const dummyMessageBus = {};
+
+    const app = createSSRApp(App, {
+        resourcesBasePath,
+        message,
+        pingButtonLabel,
+        messageBus: dummyMessageBus,
+    });
+
+    return renderToString(app);
+};
+
+export default bootstrap;